refactor : Logo 컴포넌트 생성 및 분리
